#291: moved types/ into app/
